requirements.txt
# Core
anthropic>=0.30.0
streamlit>=1.35.0
python-dotenv>=1.0.0

# E*Trade
rauth>=0.7.3
requests>=2.31.0
requests-oauthlib>=1.3.1

# Financial Data
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.26.0

# Visualisation
plotly>=5.18.0

# News / NLP
newsapi-python>=0.2.7
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Utilities
tenacity>=8.2.0
cachetools>=5.3.0
python-dateutil>=2.9.0
pytz>=2024.1
